02_SSBW/Proyecto/backend/requirements.txt
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==2.1
django-braces==1.12.0
django-cors-headers==2.1.0
django-filter==1.1.0
django-oauth-toolkit==1.0.0
django-rest-framework-mongoengine==3.3.1
django-rest-framework-social-oauth2==1.0.8
djangorestframework==3.7.7
idna==2.6
Markdown==2.6.11
mongoengine==0.15.0
oauthlib==2.0.6
PyJWT==1.5.3
pymongo==3.6.0
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
rest-social-auth==4.0.0
social-auth-app-django==3.1.0
social-auth-core==3.3.0
urllib3==1.22
djongo